tcp_stat: stop if `ss` exits with an error
Use `subprocess.check_output()` instead of `subprocess.run()`.

Note that this is an incompatible change in the behavior of
the `tcp_stat` tool: if the `ss` program becomes temporarily
unavailable, e.g. because of some kind of filesystem error or
package upgrade, `tcp_stat` will now stop immediately instead of
going on and producing another delta line when `ss` becomes
available again. It may be a better idea to add a command-line
option (e.g. `-k / --keep-going` as in make(1)) and switch back to
`subprocess.run()`, but adding a `check=True/False` parameter.

Ruff: PLW1510